fix: move restart button to settings
